# src/Ateliex.WinUI/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Ateliex.Web",
    "version": "1.0"
  },
  "paths": {
...
    "/api/ModeloRecursoAnexo/{id}": {
      "get": {
        "tags": [
          "ModeloRecursoAnexo"
        ],
        "operationId": "GetModeloRecursoAnexo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ModeloRecursoAnexo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModeloRecursoAnexo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModeloRecursoAnexo"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ModeloRecursoAnexo"
        ],
        "operationId": "PutModeloRecursoAnexo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeloRecursoAnexo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeloRecursoAnexo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModeloRecursoAnexo"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Modelo": {
        "required": [
          "id",
          "nome"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nome": {
            "maxLength": 255,
            "type": "string"
          },
          "custoDeProducao": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "recursos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModeloRecurso"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ModeloRecurso": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "modeloId": {
            "type": "integer",
            "format": "int32"
          },
          "modelo": {
            "$ref": "#/components/schemas/Modelo"
          },
          "tipoId": {
            "type": "integer",
            "format": "int32"
          },
          "tipo": {
            "$ref": "#/components/schemas/ModeloRecursoTipo"
          },
          "descricao": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "custo": {
            "type": "number",
            "format": "double"
          },
          "unidades": {
            "type": "integer",
            "format": "int32"
          },
          "custoPorUnidade": {
            "type": "number",
            "format": "double",
            "readOnly": true
          },
          "observacao": {
            "$ref": "#/components/schemas/ModeloRecursoObservacao"
          },
          "anexos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModeloRecursoAnexo"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ModeloRecursoAnexo": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "recursoId": {
            "type": "integer",
            "format": "int32"
          },
          "recurso": {
            "$ref": "#/components/schemas/ModeloRecurso"
          },
          "nome": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "arquivo": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ModeloRecursoObservacao": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "recursoId": {
            "type": "integer",
            "format": "int32"
          },
          "recurso": {
            "$ref": "#/components/schemas/ModeloRecurso"
          },
          "texto": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ModeloRecursoTipo": {
        "required": [
          "id",
          "nome"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nome": {
            "maxLength": 255,
            "type": "string"
          },
          "descricao": {
            "$ref": "#/components/schemas/ModeloRecursoTipoDescricao"
          },
          "recursos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModeloRecurso"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ModeloRecursoTipoDescricao": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "tipoId": {
            "type": "integer",
            "format": "int32"
          },
          "tipo": {
            "$ref": "#/components/schemas/ModeloRecursoTipo"
          },
          "texto": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}